Language switch refactor
