+++ hw-004/README.md
@@ -11,8 +11,15 @@ mvn clean package
 cd target
 java -Dfile.encoding=UTF-8 -jar hw-004-1.0-SNAPSHOT.jar
 ````
+или
+````
+mvn spring-boot:run -Dspring-boot.run.jvmArguments="-Dfile.encoding=UTF-8"
+````
 Использование shell:
 ````
 register - для регистрации номера зачетной книжки
 exam - для проведения экзамена
+language langCode [countryCode] - для переключения языков
+                                  пример: language en
+                                          language ru RU  
 ````
