+++ hw-004/src/main/java/ru/otus/glavs/l10n/LocalizedMessages.java
@@ -1,9 +1,6 @@
 package ru.otus.glavs.l10n;
 
-import java.util.Locale;
-
 public interface LocalizedMessages {
     String getTextMessage(String propCode);
-    void changeDefaultLocale(Locale locale) throws LocaleNotSupportedException;
-    Locale getLocale();
+
 }
