+++ hw-004/src/main/java/ru/otus/glavs/properties/LocaleNotSupportedException.java
@@ -1,4 +1,4 @@
-package ru.otus.glavs.l10n;
+package ru.otus.glavs.properties;
 
 public class LocaleNotSupportedException extends Exception {
     public LocaleNotSupportedException(String message) {
