+++ hw-004/src/main/java/ru/otus/glavs/properties/LocaleProperties.java
@@ -4,4 +4,5 @@ import java.util.Locale;
 
 public interface LocaleProperties {
     Locale getLocale();
+    void changeDefaultLocale(Locale locale) throws LocaleNotSupportedException;
 }
